Extract image layer constant snapshot writer
